+++ server/src/nest/auth/auth.module.ts
@@ -5,6 +5,7 @@ import { AuthPublicController } from './auth-public.controller';
 import { AuthController } from './auth.controller';
 import { PasskeyController } from './passkey.controller';
 import { AuthService } from './auth.service';
+import { UserProfileService } from './user-profile.service';
 import { PasskeyService } from './passkey.service';
 import { AuthMcp } from './auth.mcp';
 import { UserCleanupService } from './user-cleanup.service';
@@ -38,7 +39,7 @@ import { TripMembershipModule } from '../trip-membership/trip-membership.module'
 @Module({
   imports: [RateLimitModule, AuditModule, PermissionsModule, TripMembershipModule, MailerModule, AppConfigModule, TokensModule],
   controllers: [AuthPublicController, AuthController, PasskeyController],
-  providers: [AuthService, PasskeyService, UserCleanupService, WebauthnConfigService, AuthMcp],
+  providers: [AuthService, UserProfileService, PasskeyService, UserCleanupService, WebauthnConfigService, AuthMcp],
   exports: [AuthService, PasskeyService, UserCleanupService],
 })
 export class AuthModule {}
